refactor: Move chunk processing out of server (#277)
The goal here is to make server.cob more readable by moving most code
related to chunk processing to its own file.

To make the view distance accessible from both locations, it is moved to
the server properties copybook. At some point, the view distance will be
configurable via server.properties, but for now it is just a constant.