fix: Sync inventory after player actions (#285)
There are certain player actions that we don't fully handle yet, such as
dropping items or placing blocks in survival mode. This previously led to
desynchronization between the inventory displayed on the client and the
actual one on the server. We now send the full inventory again after any
potentially problematic action to avoid this (until there is a better way).