feat: Allocate chunk memory as needed
There is no longer a large block of memory pre-allocated for a fixed number
of chunks, but instead, there is now a high-capacity (but small) table of
pointers. `ALLOCATE` and `FREE` are used to manage the memory pointed to,
and it can be read as a chunk structure via a `BASED` data item that is
available as a copybook.

Using this method, the memory required scales with the server activity. For
instance, with no players (just the spawn chunks loaded), only a few
megabytes will be consumed, whereas many more chunks will be loaded
when players join the world. Previously, there could only be very few players,
especially if they were far apart, since the chunk capacity was a bottleneck.

Stress-test with hundreds of chunks loaded looks good, and chunks get
properly unloaded without memory leaks.