refactor: Register packet handlers on server startup
Instead of performing a packet ID to packet name lookup every time we
receive a packet, followed by expensive string comparisons, we can convert
the packet names to their IDs once when the server is starting and store the
numerical IDs with their associated program pointers for fast lookup when a
packet is received. This is currently done for the "play" client state; other
states need to be refactored to use individual handler programs first.