feat: Implement window close callbacks
This patch adds a callback system for windows, that enables implementing
specializations for different window types (such as the player inventory,
crafting table GUI, etc.). Initially, one callback type is supported: handling of
the window close event. This means we are now able to place items picked up
by the player with the mouse cursor back into their inventory when the
inventory is closed while the item is still held. Previously, the item would just
continue to be held, until the player reopened their inventory and placed it
down manually, or disconnected from the server, whereby the item would be
lost. The close callback is triggered when: 1) clients send a container_close
packet; 2) a client disconnects from the server; 3) the server is closed.