feat: Support moving items around in other inventories
Implements support for clicking on slots within the crafting table GUI, so that
items can be moved between all of the displayed slots, and this is reflected in
the player's main inventory as well. When the crafting table GUI is opened, the
initial state of all slots is sent to the player via a new "window sync" callback.
Translating slot indices to their storage locations is done via a "set slot"
callback. When the GUI is closed, crafting inputs as well as the mouse item
are moved back to the player inventory if possible.

Note that crafting is not implemented here, but this is certainly a
prerequisite :)