feat: Implement shapeless 2x2 and 3x3 crafting recipes (#314)
Recipes are loaded from the default datapack on startup. They are transformed
into a serialized format, which is also easy to do for a crafting grid inside the
player inventory or crafting table GUI, and hence allows for fast comparison
of the crafting grid against all recipes.

Quite a lot of recipes are now possible to craft, such as planks from logs or
dying wool or beds, but many other recipes are shaped and as such not
possible yet. Startup time also increases substantially due to the increase in
data that must be read and loaded, and should be optimized as future work.