feat: Support shaped crafting recipes
Follow-up to #314. This patch implements parsing of recipe files with the
type "minecraft:crafting_shaped", including very complex recipes such as the
barrel (no, really - due to there being 12 types of wood in the game, there
are 12^8 = 429,981,696 ways to craft a barrel!), along with the required
checks when crafting grid slots are changed.

This means we now support 932 out of 978 crafting recipes, with the
remaining ones needing specialized implementations, such as keeping NBT
data when coloring shulker boxes or repairing tools by combining them.

In any case, it's now possible to go from logs to planks to a crafting table, and
also towards sticks and then torches or even a cobblestone pickaxe.