feat: Add item entities (#331)
Chunks now have a linked list of entities they contain, and those entities are
sent to clients when the rest of the chunk data is sent. Entities can be created
and removed (item entities despawn automatically after 5 minutes), which
updates the linked list, and notifies players. A random UUID generator and
entity ticking as part of the gameloop are also implemented.

Together with #330, it is now possible to drop items into the world from
the player inventory or container windows! :)