feat: Parse item stack size from data generator (#334)
Knowing each item's maximum stack size, which is available as the
`minecraft:max_stack_size` data component from the data generator's
`items.json` file, we are able to merge item stacks properly when storing items
into the player's inventory.

This is currently most noticeable with crafting screens. Now, when the screen
is closed with items still present in the grid (uncrafted), they stack with any
compatible items in the player's inventory, instead of being stored into their
own slots or getting dropped into the world if this isn't possible.