feat: Consume blocks when placed in survival mode (#336)
Besides decrementing the stack size of blocks when placed, buckets are now
also replaced with their filled or empty variant when interacting with a fluid
source. This made it necessary to implement a common method for replacing
the consumed item (e.g., replacing an empty bucket with a filled bucket).
When the player's inventory is already full, the replacement item cannot be
added and must be dropped into the world. Hence, this patch also includes a
refactor for dropping item stacks originating from players.