feat: Implement crude movement physics for item entities
Item entities now have velocity, initialized depending on whether they are the
result of a broken block or thrown by the player (in the direction the player is
looking). The velocity is updated each tick and affects the item's position.
Very primitive (and wrong) collision detection is implemented for block
collisions to stop the item's movement. While incorrect, it serves as a starting
point and feels surprisingly alright most of the time.