fix: Handle "on ground" flag and properly reject out-of-order moves
Due to past refactoring work, the "ignore movement until position is confirmed"
idea got lost. Now we bring it back by separating the packet parsing from
actually adopting the new player position/rotation/flags, which is also moved
to a single method in `players.cob` that all four packet types delegate to.

Additionally, the "on ground" flag is now saved and restored in player.dat files
as well, and included in sent packets.