feat: Implement fall damage and rework damage system
The vertical distance travelled while not on the ground (as reported by the
client) is now measured, and is used to damage the player when "on ground"
becomes true again. This causes a sound effect to be played to all players.
If the player's health goes to zero, the appropriate death message is selected
based on the `minecraft:damage_type` registry.

Primitive health regeneration is also added such that there is a way to recover
from this damage besides dying.