feat!: Allow configuring level-name, default "world"
The directory name into which the world is saved was previously
hard-coded to "save". It now defaults to "world" (same as the official
Minecraft server), and is configurable via `level-name` in the
`server.properties` file.

To continue using your existing world from before this commit, set
`level-name=save` in your `server.properties`.