feat: Allocate region file structures dynamically
Previously, there would always be memory reserved for 16 region files.
This is simultaneously a large amount of space to reserve by default, and
also a relatively low limit in case of multiple players and/or a large world.
Now, the memory is allocated as needed, with a capacity of 1024 pointers
to `BASED` items.