feat: Implement save and load for entities
Most, if not all entities can be loaded from existing worlds, although
they will likely have incomplete data, as only item entities have the
logic in place for decoding type-specific data. However, this means
that dropped items will now persist across chunk unloads.

To implement this feature, the region file management is extended
to deal with files of different types - This is because both the block
data and the entity data are stored in files using the same "Anvil"
binary format, but in separate directories.