feat: Start work on a code generator (#357)
* feat: Start work on a code generator

The block loot tables, which decide which item to drop when a block is
broken, are described in highly complex JSON files with arbitrary
nesting. This makes it difficult to work with them at runtime, as we
would have to allocate dynamic memory and have lots of nested program
pointers. Instead this patch is the first step towards code generation.

It introduces a new COBOL program `codegen`, which reads the block loot
tables and generates additional COBOL source code. This code is then
compiled and linked into the main program to implement the block loot
tables.

No parsing of the JSON is done yet, as this is just a proof of concept.
All blocks still drop the item of the same name, except for stone,
which drops cobblestone.

* refactor: Make code generator template-based

Templates have the advantage of being much more readable and
can be syntax-highlighted by the IDE.