feat: Implement block loot table with code generator
Continuation of #357. This patch adds loading of the JSON data files
which describe the block loot tables. A loot table defines one or more
pools, each evaluated independently when the block is destroyed to
determine the items to drop. Pools can have conditions which must be
met for the pool to be evaluated. Further, pools can use logical
constructs, such as "alternatives" which pick the first subordinate
entry that meets all of its conditions, discarding all others.

The codegen program uses template files along with callback programs
that parse the JSON and emit the correct values in locations where the
template has placeholders (`$VARNAME$`).

At the moment, approximately 31,000 lines or 900 kiB of COBOL code is
generated. The codegen is quite efficient and takes on the order of
100 milliseconds to generate the code; however, it can take many
seconds to compile with cobc.

To test: Destroy a block such as stone, grass, or glass panes and be
amazed that these now drop the correct items (e.g., cobblestone, dirt,
and nothing, respectively).