test: Implement test framework using copybooks
Running tests is now as simple as `make test`, which properly sets the
exit code to 1 if any tests fail, and provides a summary of the test
results. This includes a listing of all failed tests that failed.

The framework keeps track of test suites, units, and cases, and ensures
that each of these contains at least 1 child. This is to enforce that
tests are properly written and include assertions.

The following copybooks are provided and used throughout the tests:

* `TEST-SUITE`: marks the beginning of a test file, such as JSON encode
* `TEST-UNIT`: starts a test unit, such as encoding a string
* `TEST-CASE`: starts a test case, such as encoding the empty string
* `TEST-ASSERT`: completes the case by evaluating a boolean expression,
    displaying either "PASS" or "FAIL" and incrementing the appropriate
    counters
* `TEST-SKIP`: explicitly marks a test case as skipped, allowing it not
    to include an assertion