feat: Implement loot table functions and number providers (#370)
This patch expands on the code generator for loot tables, adding
support for "functions" and "number providers". Functions have the
ability to modify item stacks, such as setting the count (possibly
based on some condition or a statistical distribution). Number
providers then implement these distributions, among other things.

In-game, the effect can be seen e.g. with leaves, which can now drop
between 0 and 2 sticks (previously only 0 or 1 was possible). Lapis
lazuli drops a random distribution between 4 and 9 items. Melon and
pumpkin stems follow a binomal distribution conditioned on the
growth stage of the stem.