feat: Implement `block_state_property` loot condition (#371)
This loot table condition performs a check on the mined block,
comparing its basic type to an expected value as well as possibly
comparing additional properties.

For instance, when a slab has "type=double", it should drop 2 items
instead of 1. This was previously implemented manually but is now
derived automatically from the loot table JSON.

Other affected blocks include plants that check for growth stages
to determine drops.