feat: Add codegen optimizer (#376)
With some rudimentary static analysis, we can automatically remove
redundant statements from the generated COBOL code. This reduces the
number of lines for `blocks_loot_table.cob` from 48,507 to 46,931, and
the number of bytes from 1.77 MB to 1.71 MB.

Two optimizations are implemented for now:

- Avoid `MOVE <> TO <var>` when followed by another `MOVE <> TO <var>`
- Avoid `MOVE <a> TO <b>` when preceded by `MOVE <b> TO <a>`

The optimizations are repeated until no further changes are possible.

Codegen time is basically unchanged, but the code should compile and
run slightly faster.

This patch also improves the templating code to avoid indenting empty
lines.