feat: Implement block entity allocate/serialize/deserialize
Previously, only the type ID of block entities could be stored, which
is sufficient for some blocks (that need block entities) but not
others. This patch adds a callback-based (i.e., extensible) interface
for dynamic allocation, NBT serialization, and NBT deserialization of
block entities along with changes to networking and save/load.

As a proof-of-concept, the "minecraft:sign" block entity is added,
making it possible to load signs from imported worlds, properly save
them again, and send their text to clients. Editing signs will be part
of a future patch.