refactor: Store block states by ID and update naming (#382)
Same as #380, but for block states. Using the block registry IDs to index
the block states table has the advantage of not having to loop over the
entire table to find the block states belonging to a known registry ID.

Additionally, this patch renames many related variables and programs
to hopefully improve consistency.

Some lookups are now more effort (code-wise), but should be mostly
without performance regressions, and certainly offer more paths to
optimize in the future.